client-api: Unify RtcFocusInfo and RtcTransport into one type for MSC4143/4195.
